requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
yfinance>=0.2.40
plotly>=5.20.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
apscheduler>=3.10.0
pytz>=2024.1
numpy>=1.26.0
